backend-python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
duckdb==1.1.3
pandas==2.2.3
httpx==0.28.1
psutil==6.1.0
faiss-cpu==1.9.0.post1
numpy>=1.26,<2
sentence-transformers==3.3.1
pypdf==5.1.0
pdfplumber==0.11.4
